// src/models/PanFailedPaymentResponse.model.ts
import { StatusPaymentPending } from "src/utils/ValidApplicationStatus";

export type PanFailedPaymentListResponse = {
  _id: string
  category: string
  title: string
  name: string
  dob: string
  parentName: string
  parentType: string
  adhaarNumber: string
  mobileNumber: string
  email: string
  panCardFront: string
  passportPhotoUrl: string
  signaturePhotoUrl: string
  panFormFrontPhotoUrl: string
  panFormBackPhotoUrl: string
  adhaarFrontPhotoUrl: string
  adhaarBackPhotoUrl: string
  otherDocuments: any[]
  srn: string
  distributorCode: string
  distributorId: string
  txnId: string
  uniqueTransactionId: string
  txnObjectId: string
  paymentCategory: string[]
  version: string
  panNumber: string
  totalPrice: number
  mainCategoryPrice: number
  additionalCat2: number
  additionalCat3: number
  convinienceCharges: number
  basePrice: number
  baseCatAppied: string
  refundWalletAmountApplied: number
  rewardWalletAmountApplied: number
  acknowledgementNumber: string
  acknowledgementPdf: string
  rejectionReason: string
  remark: string
  status: StatusPaymentPending
  comments: string
  applicationType: string
  appliedFrom: string
  txnStatus: string
  appliedByType: string
  appliedByUserCurrentAddress: string
  appliedById: string
  appliedOnDate: string
  appliedByName: string
  appliedByNumber: string
  appliedAsType: string
  assignedToName: string
  assignedToId: string
  assignedOnDate: string
  assignedByName: string
  assignedById: string
  verifiedByName: string
  verifiedById: string
  verifiedOnDate: string
  rejectedByName: string
  rejectedById: string
  rejectedOnDate: string
  generatedByName: string
  generatedById: string
  generatedOnDate: string
  completedByName: string
  completedById: string
  completedOnDate: string
  cancelledByName: string
  cancelledById: string
  cancelledOnDate: string
  moveToPendingByName: string
  moveToPendingById: string
  moveToPendingOnDate: string
  IsAgreedToTermsAndConditions: boolean
  retailerId: string
  retailerName: string
  retailerFirmName: string
  retailerFirmAddress: string
  retailerMobileNumber: string
  retailerPanNumber: string
  isDeleted: boolean
  serialNumber: any
  createdAt: string
  updatedAt: string
  __v: number
};